Created and implemented the WatchlistComponent that will be called in the Watchlist page
